+++ src/BecklynRadBundle.php
@@ -3,6 +3,8 @@
 namespace Becklyn\RadBundle;
 
 use Becklyn\RadBundle\Bundle\BundleExtension;
+use Becklyn\RadBundle\DependencyInjection\DoctrineExtensionsCompilerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -16,5 +18,13 @@ class BecklynRadBundle extends Bundle
     public function getContainerExtension ()
     {
         return new BundleExtension($this);
+    }
+
+
+    /**
+     */
+    public function build (ContainerBuilder $container)
+    {
+        $container->addCompilerPass(new DoctrineExtensionsCompilerPass());
     }
 }