+++ src/BecklynRadBundle.php
@@ -4,6 +4,7 @@ namespace Becklyn\RadBundle;
 
 use Becklyn\RadBundle\Bundle\BundleExtension;
 use Becklyn\RadBundle\DependencyInjection\DoctrineExtensionsCompilerPass;
+use Becklyn\RadBundle\Usages\EntityUsagesProviderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -23,8 +24,12 @@ class BecklynRadBundle extends Bundle
 
     /**
      */
-    public function build (ContainerBuilder $container)
+    public function build (ContainerBuilder $container) : void
     {
         $container->addCompilerPass(new DoctrineExtensionsCompilerPass());
+
+        $container
+            ->registerForAutoconfiguration(EntityUsagesProviderInterface::class)
+            ->addTag("entity_usages.provider");
     }
 }