+++ src/Route/DeferredRoute.php
@@ -2,12 +2,10 @@
 
 namespace Becklyn\RadBundle\Route;
 
-
 use Symfony\Component\Routing\Exception\InvalidParameterException;
 use Symfony\Component\Routing\Exception\MissingMandatoryParametersException;
 use Symfony\Component\Routing\Exception\RouteNotFoundException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-
 
 /**
  * A deferred route.