+++ CHANGELOG.md
@@ -6,6 +6,7 @@ next minor
     *   numeric functions: `RAND()`
     *   datetime functions: `WEEK()`, `YEAR()`
 *   (internal) Properly set branch alias.
+*   (feature) Add `DeferredRoute`.
 
 
 7.0.7