+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 next minor
 ==========
 
-*   (feature) Add and autoregister doctrine mysql extensions: 
+*   (feature) Added `EntityUsagesFinder`, that helps when trying to collect usages of entities.
+*   (feature) Add and autoregister doctrine mysql extensions (`doctrine` is now a required package of this bundle)
     *   string functions: `CHAR_LENGTH()`, `DATE_FORMAT()`, `IFELSE()`, `IFNULL()`, `JSON_EXTRACT()`, `LPAD()`
     *   numeric functions: `RAND()`
     *   datetime functions: `WEEK()`, `YEAR()`